Don't tell user they've failed while week is still in progress
History tooltips previously said "Failed" whenever the week total
was below the goal. However, this is misleading when the week is
not yet finished. It now says "In progress" whenever the week
hasn't ended.

Also, re-fix the tooltips not disappearing. Perhaps the result of
a 'git pull' blunder on my part?